update pto interface name



# message auto-generated for no-merge-commit merge:
!793 merge dev into master

update pto interface name

Created-by: zhangyi371
Commit-by: zhangyi371
Merged-by: turing_project1
Description: - 增加冒烟用例
- 修改pto绑定接口名
- 修改srcRepeatStride 数据类型

See merge request: cann/pypto-dev!793

+++ framework/src/interface/tileop/a2a3/dynamic/vector_dyn.h
@@ -227,7 +227,7 @@ template <typename T, unsigned DS, unsigned SS, unsigned TBS>
 TILEOP void DynTrowsumsingle_(__ubuf__ T *dst, __ubuf__ T *src, __ubuf__ T *tmp, unsigned OS0, unsigned OS1) {
     //    OS0 <= REPEAT_MAX
     uint64_t srcRepeatPerRow = static_cast<uint64_t>(OS1 * sizeof(T) / REPEAT_BYTE);
-    uint8_t srcRepeatStride = SS * sizeof(T) / BLOCK_SIZE;
+    uint16_t srcRepeatStride = SS * sizeof(T) / BLOCK_SIZE;
     constexpr unsigned nElemPerRepeat = REPEAT_BYTE / sizeof(T);
     unsigned remain = OS1 % nElemPerRepeat;
     if (srcRepeatPerRow == 1 && remain == 0) {
@@ -321,7 +321,7 @@ template <typename T, unsigned DS, unsigned SS, unsigned TBS>
 TILEOP void DynTrowmaxsingle_(__ubuf__ T *dst, __ubuf__ T *src, __ubuf__ T *tmp, unsigned OS0, unsigned OS1) {
     //    OS0 <= REPEAT_MAX
     uint64_t srcRepeatPerRow = static_cast<uint64_t>(OS1 * sizeof(T) / REPEAT_BYTE);
-    constexpr unsigned srcRepeatStride = SS * sizeof(T) / BLOCK_SIZE;
+    constexpr uint16_t srcRepeatStride = SS * sizeof(T) / BLOCK_SIZE;
     constexpr unsigned nElemPerRepeat = REPEAT_BYTE / sizeof(T);
     unsigned remain = OS1 % nElemPerRepeat;
     if (srcRepeatPerRow == 1 && OS0 <= REPEAT_MAX && remain == 0) {
@@ -403,7 +403,7 @@ template <typename T, unsigned DS, unsigned SS, unsigned TBS>
 TILEOP void DynTrowminsingle_(__ubuf__ T *dst, __ubuf__ T *src, __ubuf__ T *tmp, unsigned OS0, unsigned OS1) {
     //    OS0 <= REPEAT_MAX
     uint64_t srcRepeatPerRow = static_cast<uint64_t>(OS1 * sizeof(T) / REPEAT_BYTE);
-    constexpr unsigned srcRepeatStride = SS * sizeof(T) / BLOCK_SIZE;
+    constexpr uint16_t srcRepeatStride = SS * sizeof(T) / BLOCK_SIZE;
     constexpr unsigned nElemPerRepeat = REPEAT_BYTE / sizeof(T);
     unsigned remain = OS1 % nElemPerRepeat;
     if (srcRepeatPerRow == 1 && OS0 <= REPEAT_MAX && remain == 0) {
